+++ src/components/price.js
@@ -1,8 +1,19 @@
-export const createPriceTemplate = (tripInfo) => {
+const getTotalPrice = (tripArray) => {
+  let sum = 0;
+  tripArray.forEach((point) => {
+    sum += point.price;
+    point.offers.forEach((offer) => {
+      sum += offer.price;
+    });
+  });
+  return sum;
+};
+
+export const createPriceTemplate = (tripArray) => {
   return (
     `<p class="trip-info__cost">
       Total: &euro;&nbsp;<span class="trip-info__cost-value"
-      >${tripInfo.summaTrip}</span>
+      >${getTotalPrice(tripArray)}</span>
     </p>`
   );
 };
